Adicionando nova coluna e levantando ambiente

+++ core/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Pessoa     # importando o model Pessoa criado
 
 # Create your views home a index.
@@ -21,30 +21,36 @@ def salvar(request):
                 # IMPRIMIR: Crio o dado pela view, pego a lista toda envio para index -> 
                     # -> No index faço um loop para pegar as pessoa para imprimir em tela
 
-    vnome = request.POST.get("nome") # Pego o dado inserido do form via POST após o submit e guardei na variavel nome
-    Pessoa.objects.create(nome=vnome) # Criei no BD o novo nome da nova pessoa 
-    pessoas = Pessoa.objects.all()  # Agora evio uma lista atualizada de nomes para meu index
-    return render(request, "index.html", {"pessoas":pessoas})  # e envia para meu forme
+    if request.method == "POST":
+        vnome = request.POST.get("nome") # Pego o dado inserido do form via POST após o submit e guardei na variavel nome
+        vemail = request.POST.get("email")
+        Pessoa.objects.create(nome=vnome, email=vemail) # Criei no BD o novo nome da nova pessoa 
+        return redirect('home') # e envia para meu forme
+    else:
+        return render(request, "index.html")
 
-
-# EDITE
-def editar(request, id):   
-    #Editar recebe o ID da pessoa
-    pessoa = Pessoa.objects.get(id=id)   # Pega o ID por pessoa pelo get
+# EDIT
+def editar(request, id):
+    pessoa = get_object_or_404(Pessoa, id=id)
     return render(request, "update.html", {"pessoa": pessoa})
-
 
 # READ UPDATE
 def update(request, id):
-    vnome = request.POST.get("nome")  # nome atualizado
-    pessoa = Pessoa.objects.get(id=id) # pegar a pessoa do BD pq ela existe
-    pessoa.nome = vnome  # o novo valor que veio do banco
-    pessoa.save()
-    return redirect(home)
+    pessoa = get_object_or_404(Pessoa, id=id)
+    if request.method == "POST":
+        pessoa.nome = request.POST.get("nome")  # nome atualizado
+        pessoa.email = request.POST.get('email')  # email atualizado
+        pessoa.save()
+        return redirect('home')  # certifique-se de que 'home' é o nome correto da sua URL
+    else:
+        # Se não for um POST, mostre o formulário com os dados atuais da pessoa
+        return render(request, 'update.html', {'pessoa': pessoa})
+
 
 
 # DELETE
 def delete(request, id):
-    pessoa = Pessoa.objects.get(id=id)
-    pessoa.delete()
-    return redirect(home)
+    if request.method == "POST":
+        pessoa = get_object_or_404(Pessoa, id=id)
+        pessoa.delete()
+    return redirect('home')
